Add basic unit test framework
