+++ .github/workflows/test.yml
@@ -28,3 +28,8 @@ jobs:
     - run: npm test
       env:
         CI: true
+    - name: Upload code coverage
+      uses: codecov/codecov-action@v1
+      with:
+        file: ./coverage/coverage-final.json
+
