+++ .gitignore
@@ -1,5 +1,6 @@
 /dist/*
 /node_modules
+/coverage
 .env
 .eslintcache
 .DS_Store
