+++ src/test/tsconfig.json
@@ -1,0 +1,12 @@
+{
+	"compilerOptions": {
+		"target": "es6",
+		"module": "commonjs",
+		"incremental": true,
+		"strict": true,
+        "noEmit": true,
+	},
+	"include": [
+		"./**/*.ts"
+	],
+}
