+++ tsconfig.json
@@ -11,6 +11,6 @@
 		"src/**/*.ts", "globals.ts"
 	],
 	"exclude": [
-		"node_modules"
+		"node_modules", "src/test"
 	]
 }
